integer math

+++ main/tinybit/graphics.h
@@ -20,7 +20,7 @@ void lua_setup_draw();
 int millis();
 int random_range(int, int);
 void draw_sprite(int sourceX, int sourceY, int sourceW, int sourceH, int targetX, int targetY, int targetW, int targetH);
-void draw_sprite_rotated(int sourceX, int sourceY, int sourceW, int sourceH, int targetX, int targetY, int targetW, int targetH, float angleDegrees);
+void draw_sprite_rotated(int sourceX, int sourceY, int sourceW, int sourceH, int targetX, int targetY, int targetW, int targetH, int angleDegrees);
 void draw_rect(int x, int y, int w, int h);
 void draw_oval(int x, int y, int w, int h);
 void set_stroke(int width, int r, int g, int b, int a);
